[Silver III] Title: 주유소, Time: 372 ms, Memory: 60328 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -5,19 +5,19 @@ import java.util.Arrays;
 import java.util.StringTokenizer;
 
 class Main {
-    static int N;
-    static int sum;
-    static int[] node_Arr;
-    static int[] distance_Arr;
+    static long N;
+    static long sum;
+    static long[] node_Arr;
+    static long[] distance_Arr;
 
-    static void calculate(int now_Node){
+    static void calculate(long now_Node){
         if(now_Node == N-1){
             return;
         }
-        for(int i = now_Node; i < N; i++){
-            if(node_Arr[now_Node] > node_Arr[i] || i == N-1){
-                for(int j = now_Node; j < i; j++){
-                    sum += node_Arr[now_Node] * distance_Arr[j];
+        for(int i = Math.toIntExact(now_Node); i < N; i++){
+            if(node_Arr[Math.toIntExact(now_Node)] > node_Arr[i] || i == N-1){
+                for(int j = Math.toIntExact(now_Node); j < i; j++){
+                    sum += node_Arr[Math.toIntExact(now_Node)] * distance_Arr[j];
                 }
                 calculate(i);
                 break;
@@ -28,18 +28,18 @@ class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         N = Integer.parseInt(br.readLine());
-        node_Arr = new int[N];
-        distance_Arr = new int[N];
+        node_Arr = new long[Math.toIntExact(N)];
+        distance_Arr = new long[Math.toIntExact(N)];
 
 
         StringTokenizer st = new StringTokenizer(br.readLine());
         for(int i = 0; i < N-1; i++){
-            distance_Arr[i] = Integer.parseInt(st.nextToken());
+            distance_Arr[i] = Long.parseLong(st.nextToken());
         }
 
         StringTokenizer st2 = new StringTokenizer(br.readLine());
         for(int i = 0; i < N; i++){
-            node_Arr[i] = Integer.parseInt(st2.nextToken());
+            node_Arr[i] = Long.parseLong(st2.nextToken());
         }
 
         calculate(0);
